+++ rednotebook/util/utils.py
@@ -72,6 +72,20 @@ def setup_signal_handlers(journal):
     logging.info('Connected Signals: %s' % signals)
 
 
+def get_gtk_colors(widget):
+    """Retrieve colors of the currect GTK theme for the given widget.
+
+    The get_background_color() method is deprecated, but I couldn't find
+    a different way for retrieving the color.
+    """
+    style = widget.get_style_context()
+    bg_color = style.get_background_color(Gtk.StateFlags.NORMAL).to_string()
+    fg_color = style.get_color(Gtk.StateFlags.NORMAL).to_string()
+    logging.debug("Background color: {}".format(bg_color))
+    logging.debug("Foreground color: {}".format(fg_color))
+    return bg_color, fg_color
+
+
 def get_new_version_number():
     """
     Reads version number from website and returns None if it cannot be read